Use changeBlockMeta() in voxel-registry; add allowRecarving option
changeBlockMeta() is the opposite of getBlockMeta(), returns a block
index for a given base index + metadata. It can be used either on a base
index, or an index which already includes (non-zero) metadata; i.e.,
changing the existing metadata. Logically it only makes sense to carve a
pumpkin once, so allowRecarving is off by default, but this is a
gameplay design restriction decision not a technical limitation, so you
can set allowRecarving:true in opts if you want to allow it.

Requires deathcap/voxel-registry@5af02e942999bec898abe88f9332fa1488a4756d